+++ src/lib/holocene/tooltip.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Copyable from '$lib/components/copyable.svelte';
+  import Copyable from '$lib/holocene/copyable.svelte';
   import Icon from '$lib/holocene/icon/icon.svelte';
   import type { IconName } from '$lib/holocene/icon/paths';
   import type { Only } from '$lib/types/global';
